refactor: migrate FadeForwardsPageTransitionsBuilder to widgets
