+++ examples/api/test/widgets/page_transitions_builder/page_transitions_builder.1_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:flutter_api_samples/material/page_transitions_theme/page_transitions_theme.3.dart'
+import 'package:flutter_api_samples/widgets/page_transitions_builder/page_transitions_builder.1.dart'
     as example;
 import 'package:flutter_test/flutter_test.dart';
